fix things

+++ backend/entries/models.py
@@ -209,7 +209,7 @@ class SalesMonthlyTarget(models.Model):
     )
     year = models.PositiveIntegerField()
     month = models.PositiveIntegerField()
-    premium_target = models.DecimalField(max_digits=15, decimal_places=2, null=True, blank=True)
+    premium_target = models.PositiveIntegerField(null=True, blank=True)
     clients_assigned = models.PositiveIntegerField(null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
